🎨 move pagination string calc into pagination component 🐛 add row count formatting to query editor top message to align with pagination
